code/category_management/requirements.txt
python-dev-tools
pandas
psycopg2
PyYAML
pyspark